requirements-enhanced.txt
# Enhanced Requirements for Google Cloud and Mobile App Development
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7

# Performance and Security
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-CORS==4.0.0
redis==5.0.1

# Database drivers
psycopg2-binary==2.9.9  # PostgreSQL
PyMySQL==1.1.0         # MySQL/MariaDB (Google Cloud SQL)
sqlalchemy==2.0.23

# Google Cloud Platform
google-cloud-sql-python-connector==1.4.3
google-cloud-storage==2.12.0
google-cloud-logging==3.8.0
google-cloud-monitoring==2.16.0
google-auth==2.23.4

# Data processing and utilities
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9

# QR Code generation
qrcode[pil]==7.4.2
Pillow==10.1.0

# URL handling and security
shortuuid==1.0.11
itsdangerous==2.1.2
bcrypt==4.1.2

# Payment processing
paypalrestsdk==1.13.3

# Email functionality
smtplib-ssl==1.0.4

# Environment management
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development dependencies
black==23.11.0
flake8==6.1.0
pre-commit==3.6.0

# Mobile API dependencies
marshmallow==3.20.1      # API serialization
flask-jwt-extended==4.6.0  # JWT authentication
celery==5.3.4             # Background tasks
gunicorn==21.2.0          # WSGI server

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Additional security
cryptography==41.0.8
secrets==1.0.1

# Production server
gevent==23.9.1
eventlet==0.33.3

# Google Cloud Run compatibility
cloudsql-proxy==2.7.2

# Mobile app development support
pyjwt==2.8.0
requests==2.31.0
urllib3==2.1.0

# API documentation
flask-restx==1.3.0
apispec==6.3.0

# Rate limiting and throttling
slowapi==0.1.9

# Load testing and performance monitoring
locust==2.17.0

# Backup and migration tools
alembic==1.13.1
flask-migrate==4.0.5